xe: jit: refactor reorder interface
