+++ src/gpu/intel/jit/ir/primitive_plan.cpp
@@ -129,9 +129,8 @@ status_t primitive_init_plan_t::add_reorder_kernel(
         }
     }
     exec_config_t exec_cfg(hw_t(engine), regs_, simd_);
-    kernel_info.set_nd_range(
-            reorder_kernel_t<>::nd_range(exec_cfg, src.layout, dst.layout));
     reorder_config_t cfg(exec_cfg, src.layout, dst.layout);
+    kernel_info.set_nd_range(cfg.nd_range());
     auto kernel = make_kernel<reorder_kernel_t>(primitive,
             /*register_kernel=*/true, engine, cfg, "reorder", kernel_info,
             dpas_);
