+++ src/gpu/intel/jit/reorder/reorder_kernel.hpp
@@ -60,11 +60,6 @@ public:
 
         generate_epilogue();
     }
-
-    static compute::nd_range_t nd_range(const exec_config_t &exec_cfg,
-            const layout_t &src, const layout_t &dst) {
-        return reorder_ir_builder_t::nd_range(exec_cfg, src, dst);
-    }
 };
 
 } // namespace jit
