+++ src/gpu/intel/jit/reorder/tiler.hpp
@@ -1,0 +1,40 @@
+/*******************************************************************************
+* Copyright 2025 Intel Corporation
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*******************************************************************************/
+
+#ifndef GPU_INTEL_JIT_REORDER_TILER_HPP
+#define GPU_INTEL_JIT_REORDER_TILER_HPP
+
+#include <array>
+
+#include "gpu/intel/jit/ir/tensor.hpp"
+
+namespace dnnl {
+namespace impl {
+namespace gpu {
+namespace intel {
+namespace jit {
+namespace reorder {
+
+std::vector<tensor_t> tiles(const hw_t &hw, layout_t a, layout_t b);
+
+} // namespace reorder
+} // namespace jit
+} // namespace intel
+} // namespace gpu
+} // namespace impl
+} // namespace dnnl
+
+#endif
